refactor:fixx all the comments
