+++ src/components/GoodsList/GoodsList.jsx
@@ -1,0 +1,11 @@
+const GoodsList = ({ goods }) => (
+  <ul>
+    {goods.map(el => (
+      <li data-cy="Good" key={el}>
+        {el}
+      </li>
+    ))}
+  </ul>
+);
+
+export default GoodsList;
