+++ src/utils/constants.js
@@ -1,0 +1,3 @@
+export const SORT_BY_CHAR = 'SORT_BY_CHAR';
+export const SORT_BY_LENGTH = 'SORT_BY_LENGTH';
+export const SORT_BY_DEFAULT = '';
